+++ shared-web/components/Menu/Menu.tsx
@@ -13,13 +13,14 @@ type MenuProps = {
   trigger: ComponentChildren;
   items: MenuItem[];
   class?: string;
+  placement?: 'bottom' | 'top';
 };
 
-export function Menu({ trigger, items, class: className }: MenuProps) {
+export function Menu({ trigger, items, class: className, placement = 'bottom' }: MenuProps) {
   const [open, setOpen] = useState(false);
   const ref = useRef<HTMLDivElement>(null);
   const dropdownRef = useRef<HTMLDivElement>(null);
-  const [pos, setPos] = useState<{ top: number; right: number } | null>(null);
+  const [pos, setPos] = useState<{ top?: number; bottom?: number; right: number } | null>(null);
 
   useEffect(() => {
     if (!open) return;
@@ -28,7 +29,14 @@ export function Menu({ trigger, items, class: className }: MenuProps) {
       const el = ref.current;
       if (!el) return;
       const rect = el.getBoundingClientRect();
-      setPos({ top: rect.bottom, right: window.innerWidth - rect.right });
+      if (placement === 'top') {
+        setPos({
+          bottom: window.innerHeight - rect.top + 4,
+          right: window.innerWidth - rect.right,
+        });
+      } else {
+        setPos({ top: rect.bottom, right: window.innerWidth - rect.right });
+      }
     }
     updatePosition();
 
@@ -58,9 +66,11 @@ export function Menu({ trigger, items, class: className }: MenuProps) {
         createPortal(
           <div
             ref={dropdownRef}
-            class="vi-menu__dropdown"
+            class={['vi-menu__dropdown', placement === 'top' && 'vi-menu__dropdown--top']
+              .filter(Boolean)
+              .join(' ')}
             role="menu"
-            style={{ top: pos.top, right: pos.right }}
+            style={{ top: pos.top, bottom: pos.bottom, right: pos.right }}
           >
             {items.map((item, i) =>
               item.href ? (
