+++ web/src/pages/Devices/style.css
@@ -1,0 +1,76 @@
+.dashboard {
+  width: 100%;
+  max-width: 900px;
+  padding: var(--space-6) var(--space-4);
+  text-align: left;
+}
+
+.dashboard-section {
+  margin-bottom: 2.5rem;
+
+  & > h2 {
+    margin: 0 0 var(--space-4);
+  }
+
+  & h2 {
+    font-size: 1.1rem;
+    font-weight: 600;
+    color: var(--text-muted);
+  }
+}
+
+.dashboard-section-header {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  gap: var(--space-3);
+  flex-wrap: wrap;
+  margin-bottom: var(--space-4);
+
+  & h2 {
+    margin: 0;
+  }
+}
+
+.dashboard-section-header > * {
+  min-width: 0;
+}
+
+/* Device dialogs */
+
+.invite-desc {
+  font-size: var(--text-sm);
+  color: var(--text-muted);
+  margin: 0 0 var(--space-3);
+}
+
+.device-setup-dialog {
+  max-width: 460px;
+}
+
+.device-setup-steps {
+  margin: 0 0 var(--space-4);
+  padding-left: 1.25rem;
+  display: flex;
+  flex-direction: column;
+  gap: var(--space-3);
+  font-size: var(--text-base);
+}
+
+.device-setup-steps::marker {
+  font-size: 1.05rem;
+  font-weight: 700;
+}
+
+.device-setup-steps li {
+  color: var(--text);
+  font-size: var(--text-base);
+  line-height: var(--line-normal);
+  padding-left: 0.25rem;
+}
+
+.device-setup-step-label {
+  display: block;
+  font-weight: 600;
+  font-size: 1.02rem;
+}
