+++ web/src/pages/Logs/layout.css
@@ -21,173 +21,11 @@
   display: flex;
   width: 100%;
   min-height: 0;
-  gap: 1.75rem;
   text-align: left;
-
-  @media (max-width: 600px) {
-    flex-direction: column;
-    min-height: auto;
-  }
 }
 
 .logs-page--gallery-fullscreen .logs-layout {
   gap: 0;
-}
-
-.logs-sidebar {
-  width: 200px;
-  flex-shrink: 0;
-  min-width: 0;
-  min-height: 0;
-  overflow-y: auto;
-  overflow-x: hidden;
-  overscroll-behavior: contain;
-  scrollbar-gutter: stable;
-  padding-right: 0.25rem;
-
-  & h2 {
-    margin: 0 0 0.75rem;
-    font-size: 0.75rem;
-    font-weight: 700;
-    color: var(--text-muted);
-  }
-
-  @media (max-width: 900px) {
-    position: fixed;
-    top: 0;
-    left: 0;
-    bottom: 0;
-    z-index: 50;
-    width: min(18rem, calc(100vw - 2.75rem));
-    padding: 0.75rem;
-    background: var(--surface);
-    border-right: 1px solid var(--border);
-    overflow-y: auto;
-    transform: translateX(-100%);
-    visibility: hidden;
-    pointer-events: none;
-    transition: transform 0.18s ease;
-
-    &.is-open {
-      transform: translateX(0);
-      visibility: visible;
-      pointer-events: auto;
-      box-shadow: var(--shadow-md);
-    }
-  }
-}
-
-.logs-sidebar .logs-sidebar-header {
-  display: none;
-
-  @media (max-width: 900px) {
-    display: flex;
-    margin-bottom: 0.75rem;
-  }
-}
-
-button.logs-sidebar-close {
-  display: none;
-
-  @media (max-width: 900px) {
-    display: inline-flex;
-  }
-}
-
-.logs-sidebar-group {
-  margin-top: 1rem;
-  min-width: 0;
-}
-
-.logs-sidebar-group-label {
-  margin: 0 0 0.3rem 0.65rem;
-  font-size: 0.72rem;
-  font-weight: 700;
-  color: var(--text-muted);
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-}
-
-.logs-device-list {
-  list-style: none;
-  margin: 0;
-  padding: 0;
-  display: flex;
-  flex-direction: column;
-  gap: 2px;
-}
-
-.logs-device-list li {
-  min-width: 0;
-  width: 100%;
-}
-
-.logs-device-button {
-  width: 100%;
-  max-width: 100%;
-  min-width: 0;
-  background: transparent;
-  border: none;
-  border-radius: 6px;
-  padding: 0.45rem 0.65rem;
-  font-size: 0.875rem;
-  cursor: pointer;
-  color: var(--text);
-  text-align: left;
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-  overflow: hidden;
-  transition: background 0.12s;
-
-  &:hover {
-    background: var(--bg-subtle);
-  }
-
-  &.is-active {
-    background: var(--accent);
-    color: white; /* literal: text on accent button */
-
-    & .logs-status-dot--online,
-    & .logs-status-dot--offline {
-      background: white; /* literal: dot on accent button */
-    }
-  }
-}
-
-.logs-device-button-label {
-  min-width: 0;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-}
-
-.logs-status-dot {
-  width: 7px;
-  height: 7px;
-  border-radius: 50%;
-  flex-shrink: 0;
-}
-
-.logs-status-dot--online {
-  background: var(--success);
-}
-
-.logs-status-dot--offline {
-  background: var(--text-muted);
-}
-
-.logs-device-button-group {
-  color: var(--text-muted);
-  font-weight: 700;
-}
-
-.logs-sidebar-loading {
-  font-size: 0.85rem;
-  color: var(--text-muted);
-  padding: 0.4rem 0.65rem;
-  margin-left: 0.99rem;
 }
 
 .logs-main {
@@ -223,21 +61,6 @@ button.logs-sidebar-close {
   flex-wrap: wrap;
 }
 
-button.logs-sidebar-toggle {
-  display: none;
-  align-items: center;
-  gap: 0.45rem;
-
-  @media (max-width: 900px) {
-    display: inline-flex;
-  }
-
-  & svg {
-    width: 1rem;
-    height: 1rem;
-  }
-}
-
 .logs-header-view-controls {
   display: flex;
   align-items: center;
@@ -251,63 +74,19 @@ button.logs-fullscreen-btn {
   }
 }
 
-button.logs-sidebar-backdrop,
-button.logs-sidebar-backdrop.is-open {
-  display: none;
-  opacity: 0;
-  pointer-events: none;
-}
-
-@media (max-width: 900px) {
-  button.logs-sidebar-backdrop.is-open {
-    display: block;
-    opacity: 1;
-    pointer-events: auto;
-  }
-}
-
 @media (max-width: 600px) {
   .logs-header {
     margin-bottom: 1rem;
   }
 
-  .logs-header-actions {
-    display: grid;
-    grid-template-columns: auto minmax(0, 1fr);
-    grid-template-areas:
-      'toggle  view-controls'
-      'filters filters';
-    align-items: center;
-    width: 100%;
-    row-gap: 0.5rem;
-    column-gap: 0.6rem;
-  }
-
-  button.logs-sidebar-toggle {
-    grid-area: toggle;
-    justify-self: start;
-  }
-
   .logs-header-view-controls {
-    grid-area: view-controls;
-    justify-self: end;
-    margin-left: 0;
+    margin-left: auto;
     min-width: 0;
   }
 
   .logs-filter-section {
-    grid-area: filters;
+    width: 100%;
     min-width: 0;
-  }
-}
-
-@media (max-width: 420px) {
-  button.logs-sidebar-toggle span {
-    display: none;
-  }
-
-  button.logs-sidebar-toggle {
-    padding: 0 0.55rem;
   }
 }
