+++ web/src/pages/Partners/style.css
@@ -44,37 +44,6 @@
   margin: 0 0 var(--space-3);
 }
 
-.device-setup-dialog {
-  max-width: 460px;
-}
-
-.device-setup-steps {
-  margin: 0 0 var(--space-4);
-  padding-left: 1.25rem;
-  display: flex;
-  flex-direction: column;
-  gap: var(--space-3);
-  font-size: var(--text-base);
-}
-
-.device-setup-steps::marker {
-  font-size: 1.05rem;
-  font-weight: 700;
-}
-
-.device-setup-steps li {
-  color: var(--text);
-  font-size: var(--text-base);
-  line-height: var(--line-normal);
-  padding-left: 0.25rem;
-}
-
-.device-setup-step-label {
-  display: block;
-  font-weight: 600;
-  font-size: 1.02rem;
-}
-
 .partners-panel {
   display: flex;
   flex-direction: column;
