+++ web/src/style.css
@@ -1,5 +1,5 @@
 @import '@virtueinitiative/shared-web/base.css';
 @import './styles/app-shell.css';
 @import './styles/drawer.css';
-@import './styles/header.css';
+@import './styles/sidebar.css';
 @import './styles/dialog.css';
