--- web/src/styles/header.css
@@ -1,156 +1,0 @@
-.app-header {
-  position: relative;
-  z-index: 60;
-  flex-shrink: 0;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  gap: 0.75rem 1rem;
-  background-color: var(--surface);
-  border-bottom: 1px solid var(--border);
-  padding: 0.5rem 1.25rem;
-}
-
-.app-header-brand-row {
-  min-height: 2.25rem;
-  display: flex;
-  align-items: center;
-  flex: 1 1 auto;
-  min-width: 0;
-}
-
-.app-header-brand {
-  color: var(--text);
-  font-weight: 700;
-  font-size: 1.1rem;
-  letter-spacing: 0.02em;
-  line-height: 1.1;
-
-  a& {
-    text-decoration: none;
-    background: none;
-
-    &:hover {
-      color: var(--text);
-    }
-  }
-}
-
-.app-header-toolbar {
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-  flex-shrink: 0;
-}
-
-.app-header-mobile-actions,
-.app-header-mobile-menu {
-  display: none;
-}
-
-.app-header-nav-group,
-.app-header-action-group {
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-}
-
-.app-header-action-group .vi-icon-btn {
-  width: 2.25rem;
-  min-width: 2.25rem;
-  height: 2.25rem;
-}
-
-.app-header-nav-group .vi-btn--flat.is-active {
-  color: var(--accent);
-  font-weight: 700;
-}
-
-@media (max-width: 720px) {
-  .app-header {
-    align-items: center;
-  }
-
-  .app-header-brand-row {
-    min-width: 0;
-  }
-
-  .app-header-toolbar {
-    display: none;
-  }
-
-  .app-header-mobile-actions {
-    display: flex;
-    align-items: center;
-    gap: 0.5rem;
-    flex-shrink: 0;
-  }
-
-  .app-header-mobile-menu {
-    display: flex;
-    position: fixed;
-    top: 0;
-    right: 0;
-    bottom: 0;
-    z-index: 50;
-    width: min(18rem, calc(100vw - 2.75rem));
-    padding: 0.75rem 0.75rem 1rem;
-    border-left: 1px solid var(--border);
-    background: var(--surface);
-    flex-direction: column;
-    align-items: stretch;
-    gap: 0.25rem;
-    transform: translateX(100%);
-    visibility: hidden;
-    pointer-events: none;
-    transition: transform 0.18s ease;
-  }
-
-  .app-header-mobile-menu-header {
-    margin-bottom: 0.5rem;
-  }
-
-  .app-header-mobile-menu-title {
-    font-size: 1rem;
-    font-weight: 700;
-    color: var(--text);
-    text-align: left;
-  }
-
-  .app-header-mobile-menu.is-open {
-    transform: translateX(0);
-    visibility: visible;
-    pointer-events: auto;
-    box-shadow: var(--shadow-md);
-  }
-
-  .app-header-mobile-menu .vi-btn {
-    justify-content: flex-start;
-    width: 100%;
-    min-height: 2.6rem;
-    font-size: 0.95rem;
-  }
-
-  .app-header-mobile-menu .vi-btn--flat {
-    border-color: transparent;
-  }
-
-  .app-header-mobile-menu .vi-btn--flat.is-active {
-    color: var(--accent);
-    font-weight: 700;
-  }
-
-  .app-header-mobile-menu .app-header-mobile-logout {
-    margin-top: 0.5rem;
-    width: auto;
-    min-width: 8.5rem;
-    align-self: center;
-    justify-content: center;
-  }
-}
-
-@media (max-width: 420px) {
-  .app-header-mobile-menu {
-    width: min(17rem, calc(100vw - 1.5rem));
-  }
-}
